Merged PR 381264: [React Wrapper]: Clean the Event Handler Map during Unmounting
**Issue:**  https://github.com/microsoft/powerbi-client-react/issues/79

**Problem:** With Strict Mode in React 18, React will simulate mounting, unmounting and remounting the component in development mode. [React 18 Strict Mode behaviours](https://react.dev/blog/2022/03/29/react-v18#new-strict-mode-behaviors)

During the first mount Event Handlers will be set and we store the eventHandlers in a variable `prevEventHandlerMapString` to avoid resetting the eventHandlers again if the same map is passed.

After the first mount, it will unmount, and we are not cleaning the `prevEventHandlerMapString` here.
When the component gets remounted again, since there is no change in eventHandlers it will return instead of setting the eventHandlers.

**Resolution:**

Clean the `prevEventHandlerMapString` during component gets unmounted to initial empty string.

Related work items: #1069474, #1072700, #1073594